AllExerciseRecord -> MyExerciseRecord 이름 변경, 해당 클래스 내 하위 클래스 다른 파일로 이동
