+++ fitty_frontend_app/lib/data/all_exercise_record.dart
@@ -1,12 +1,13 @@
 import 'dart:convert';
+import 'package:fitty_frontend_app/data/class/day_exercise_record.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-class AllExerciseRecord extends ChangeNotifier {
+class MyExerciseRecord extends ChangeNotifier {
   Map<String, DayExerciseRecord> dateTimeToDayExerciseRecordMap = {};
   SharedPreferences prefs;
 
-  AllExerciseRecord(this.prefs) {
+  MyExerciseRecord(this.prefs) {
     var encodedData = prefs.getString('dateTimeToDayExerciseRecordMap');
     if (encodedData != null) {
       var decodedData = jsonDecode(encodedData);
@@ -41,49 +42,4 @@ class AllExerciseRecord extends ChangeNotifier {
         jsonEncode(dateTimeToDayExerciseRecordMap));
     notifyListeners();
   }
-}
-
-class DayExerciseRecord {
-  List<OneExerciseRecord> oneExerciseRecords = [];
-  // todo : 시간 등등 기록
-  DayExerciseRecord();
-  DayExerciseRecord.fromJson(Map<String, dynamic> json)
-      : oneExerciseRecords = List<OneExerciseRecord>.from(
-            json['oneExerciseRecords']
-                .map((value) => OneExerciseRecord.fromJson(value)));
-  Map<String, dynamic> toJson() {
-    return {
-      'oneExerciseRecords': oneExerciseRecords,
-    };
-  }
-}
-
-class OneExerciseRecord {
-  String exerciseName = "";
-  List<OneSetRecord> oneSetRecords = [];
-  OneExerciseRecord();
-  OneExerciseRecord.fromJson(Map<String, dynamic> json)
-      : exerciseName = json['exerciseName'],
-        oneSetRecords = List<OneSetRecord>.from(
-            json['oneSetRecords'].map((value) => OneSetRecord.fromJson(value)));
-  Map<String, dynamic> toJson() {
-    return {
-      'exerciseName': exerciseName,
-      'oneSetRecords': oneSetRecords,
-    };
-  }
-  // todo : 얘도 시간 등등 기록
-}
-
-class OneSetRecord {
-  int weight = 0;
-  int reps = 0;
-  OneSetRecord();
-  OneSetRecord.fromJson(Map<String, dynamic> json)
-      : weight = json['weight'],
-        reps = json['reps'];
-  Map<String, dynamic> toJson() => {
-        'weight': weight,
-        'reps': reps,
-      };
 }
