+++ fitty_frontend_app/lib/data/class/one_set_record.dart
@@ -1,0 +1,12 @@
+class OneSetRecord {
+  int weight = 0;
+  int reps = 0;
+  OneSetRecord();
+  OneSetRecord.fromJson(Map<String, dynamic> json)
+      : weight = json['weight'],
+        reps = json['reps'];
+  Map<String, dynamic> toJson() => {
+        'weight': weight,
+        'reps': reps,
+      };
+}
