+++ fitty_frontend_app/lib/page/exercise/daily_routine_page.dart
@@ -1,4 +1,5 @@
 import 'package:fitty_frontend_app/data/all_exercise_record.dart';
+import 'package:fitty_frontend_app/data/class/one_exercise_record.dart';
 import 'package:fitty_frontend_app/page/exercise/exercise_record_list_page.dart';
 import 'package:fitty_frontend_app/widget/exercise/one_exercise_record_widget.dart';
 import 'package:flutter/material.dart';
@@ -23,7 +24,7 @@ class DailyRoutinePage extends StatelessWidget {
           title: Text(title),
         ),
         body: SingleChildScrollView(
-          child: Center(child: Consumer<AllExerciseRecord>(
+          child: Center(child: Consumer<MyExerciseRecord>(
             builder: (context, allExerciseRecord, child) {
               var selectedDayExerciseRecord =
                   allExerciseRecord.getDayExerciseRecord(selectedDay);
