+++ fitty_frontend_app/lib/page/exercise/exercise_record_list_page.dart
@@ -1,5 +1,7 @@
 import 'dart:convert';
 import 'package:fitty_frontend_app/data/all_exercise_record.dart';
+import 'package:fitty_frontend_app/data/class/day_exercise_record.dart';
+import 'package:fitty_frontend_app/data/class/one_exercise_record.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:intl/intl.dart';
