+++ fitty_frontend_app/lib/utility/exercise_data_processer.dart
@@ -1,9 +1,10 @@
 import 'package:fitty_frontend_app/data/all_exercise_record.dart';
+import 'package:fitty_frontend_app/data/class/one_exercise_record.dart';
 
 class ExerciseDataProcesser {
   static Map<String, OneExerciseRecord>
       getOneExerciseRecordsOfAllDateFromAllExerciseRecord(
-          AllExerciseRecord allExerciseRecord, String exerciseName) {
+          MyExerciseRecord allExerciseRecord, String exerciseName) {
     Map<String, OneExerciseRecord> oneExerciseRecordsOfAllDate = {};
     var recordEntries = allExerciseRecord.recordExistingEntries;
     for (var recordEntry in recordEntries) {
