+++ fitty_frontend_app/lib/widget/exercise/one_exercise_records_of_all_date_widget.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 import 'package:fitty_frontend_app/data/all_exercise_record.dart';
+import 'package:fitty_frontend_app/data/class/one_exercise_record.dart';
 import 'package:fitty_frontend_app/utility/exercise_data_processer.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -27,7 +28,7 @@ class _OneExerciseRecordsOfAllDateWidgetState
   @override
   void initState() {
     super.initState();
-    var allExerciseRecord = context.read<AllExerciseRecord>();
+    var allExerciseRecord = context.read<MyExerciseRecord>();
     var oneExerciseRecordsOfAllDate = ExerciseDataProcesser
         .getOneExerciseRecordsOfAllDateFromAllExerciseRecord(
             allExerciseRecord, widget.exersiseName);
