+++ fitty_frontend_app/lib/widget/exercise/one_set_record_widget.dart
@@ -1,4 +1,5 @@
 import 'package:fitty_frontend_app/data/all_exercise_record.dart';
+import 'package:fitty_frontend_app/data/class/one_set_record.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
